add custom msg

+++ models/token.go
@@ -2,5 +2,5 @@ package models
 
 type Token struct {
 	AccessToken string `json:"access_token"`
-	ExpiresIn int16 `json:"expires_in"`
+	ExpiresIn int `json:"expires_in"`
 }
